telegram_bot/requirements.txt
python-telegram-bot==21.1
langchain==0.2.6
langchain-community==0.2.6
chromadb==0.4.22
sentence-transformers==2.2.2
huggingface-hub==0.23.0
pymongo==4.6.1
python-dotenv==1.0.0
# For Ollama integration
langchain-ollama==0.1.0
